pyproject.toml
[project]
name = "simulationlab"
version = "0.1.0"
description = "CLI for composing simulation environments, running agents, and evaluating tasks"
readme = { file = "README.md", content-type = "text/markdown" }
license = { file = "LICENSE" }
authors = [
    { name = "Collinear AI" },
]
requires-python = ">=3.13"
dependencies = [
    "click>=8.0",
    "litellm>=1.80.0",
    "openai>=1.0.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "questionary>=2.0",
    "requests>=2.28",
    "rich>=13.0",
    "mcp>=1.0.0",
    "tomli-w>=1.0",
]
classifiers = [
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.13",
]

[project.urls]
Documentation = "https://docs.collinear.ai"
Homepage = "https://github.com/collinear-ai/simlab"
Issues = "https://github.com/collinear-ai/simlab/issues"
Repository = "https://github.com/collinear-ai/simlab"
Changelog = "https://github.com/collinear-ai/simlab/blob/main/CHANGELOG.md"

[project.optional-dependencies]
npc = [
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
]
daytona = [
    "daytona>=0.149.0",
]
langchain = [
    "langchain-core>=0.3.0",
    "langgraph>=0.3.0",
]
openai-agents = [
    "openai-agents>=0.9.0",
]

[project.scripts]
simlab = "simlab.cli.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/simlab"]

[tool.hatch.build.targets.sdist]
exclude = [
    "/.github",
    "/.venv",
    "/assets",
    "/examples",
    "/taskgen",
    "/tests",
    "/uv.lock",
]

[dependency-groups]
dev = [
    "build>=1.3.0",
    "mypy>=1.17.1",
    "pytest>=9.0.2",
    "ruff>=0.12.11",
    "twine>=6.1.0",
    "types-pyyaml>=6.0.12.20250915",
    "types-requests>=2.32.4.20260107",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.mypy]
explicit_package_bases = true
